[http] Move classes under 'internal' pkg as nothing should be exported
