+++ addons/binding/org.openhab.binding.http/src/main/java/org/openhab/binding/http/internal/model/Transform.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.http.model;
+package org.openhab.binding.http.internal.model;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
 import org.eclipse.smarthome.core.transform.TransformationException;
